split main.go into three files
We were already at 600+ lines of code in main.go, which isn't good.
